fixed and added pagination.

+++ Models/SeizureFormReturn.cs
@@ -1,5 +1,5 @@
 public class SeizureFormReturn
 {
-    public int PageCount { get; set; }
+    public double PageCount { get; set; }
     public List<List<SeizureFormDto>> Seizures { get; set; }
 }
